update
i can't put any more hours on this, I will present the unit test I was
able to make long ago and due to us picking up the work of the two kids
we kicked out, the workload and the refractoring has been overloaded.  i
dont think we need a present.
just fixed the login area again, it appears to have been broken..still
is in reality but I will swing by it during prensetantion.